Implement common logic for lapack syevd/heevd
