+++ dpnp/backend/extensions/lapack/syevd.hpp
@@ -1,5 +1,5 @@
 //*****************************************************************************
-// Copyright (c) 2023-2024, Intel Corporation
+// Copyright (c) 2024, Intel Corporation
 // All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without
@@ -25,10 +25,7 @@
 
 #pragma once
 
-#include <CL/sycl.hpp>
-#include <oneapi/mkl.hpp>
-
-#include <dpctl4pybind11.hpp>
+#include "evd_common.hpp"
 
 namespace dpnp
 {
@@ -38,15 +35,97 @@ namespace ext
 {
 namespace lapack
 {
-extern std::pair<sycl::event, sycl::event>
-    syevd(sycl::queue exec_q,
-          const std::int8_t jobz,
-          const std::int8_t upper_lower,
-          dpctl::tensor::usm_ndarray eig_vecs,
-          dpctl::tensor::usm_ndarray eig_vals,
-          const std::vector<sycl::event> &depends = {});
+namespace mkl_lapack = oneapi::mkl::lapack;
+namespace type_utils = dpctl::tensor::type_utils;
 
-extern void init_syevd_dispatch_table(void);
+template <typename T, typename RealT>
+static sycl::event syevd_impl(sycl::queue &exec_q,
+                              const oneapi::mkl::job jobz,
+                              const oneapi::mkl::uplo upper_lower,
+                              const std::int64_t n,
+                              char *in_a,
+                              char *out_w,
+                              std::vector<sycl::event> &host_task_events,
+                              const std::vector<sycl::event> &depends)
+{
+    type_utils::validate_type_for_device<T>(exec_q);
+    type_utils::validate_type_for_device<RealT>(exec_q);
+
+    T *a = reinterpret_cast<T *>(in_a);
+    RealT *w = reinterpret_cast<T *>(out_w);
+
+    const std::int64_t lda = std::max<size_t>(1UL, n);
+    const std::int64_t scratchpad_size =
+        mkl_lapack::syevd_scratchpad_size<T>(exec_q, jobz, upper_lower, n, lda);
+    T *scratchpad = nullptr;
+
+    std::stringstream error_msg;
+    std::int64_t info = 0;
+
+    sycl::event syevd_event;
+    try {
+        scratchpad = sycl::malloc_device<T>(scratchpad_size, exec_q);
+
+        syevd_event = mkl_lapack::syevd(
+            exec_q,
+            jobz, // 'jobz == job::vec' means eigenvalues and eigenvectors are
+                  // computed.
+            upper_lower, // 'upper_lower == job::upper' means the upper
+                         // triangular part of A, or the lower triangular
+                         // otherwise
+            n,           // The order of the matrix A (0 <= n)
+            a, // Pointer to A, size (lda, *), where the 2nd dimension, must be
+               // at least max(1, n) If 'jobz == job::vec', then on exit it will
+               // contain the eigenvectors of A
+            lda, // The leading dimension of a, must be at least max(1, n)
+            w,   // Pointer to array of size at least n, it will contain the
+                 // eigenvalues of A in ascending order
+            scratchpad, // Pointer to scratchpad memory to be used by MKL
+                        // routine for storing intermediate results
+            scratchpad_size, depends);
+    } catch (mkl_lapack::exception const &e) {
+        error_msg
+            << "Unexpected MKL exception caught during syevd() call:\nreason: "
+            << e.what() << "\ninfo: " << e.info();
+        info = e.info();
+    } catch (sycl::exception const &e) {
+        error_msg << "Unexpected SYCL exception caught during syevd() call:\n"
+                  << e.what();
+        info = -1;
+    }
+
+    if (info != 0) // an unexpected error occurs
+    {
+        if (scratchpad != nullptr) {
+            sycl::free(scratchpad, exec_q);
+        }
+        throw std::runtime_error(error_msg.str());
+    }
+
+    sycl::event clean_up_event = exec_q.submit([&](sycl::handler &cgh) {
+        cgh.depends_on(syevd_event);
+        auto ctx = exec_q.get_context();
+        cgh.host_task([ctx, scratchpad]() { sycl::free(scratchpad, ctx); });
+    });
+
+    host_task_events.push_back(clean_up_event);
+    return syevd_event;
+}
+
+template <typename fnT, typename T, typename RealT>
+struct SyevdContigFactory
+{
+    fnT get()
+    {
+        if constexpr (types::SyevdTypePairSupportFactory<T, RealT>::is_defined)
+        {
+            return syevd_impl<T, RealT>;
+        }
+        else {
+            return nullptr;
+        }
+    }
+};
 } // namespace lapack
 } // namespace ext
 } // namespace backend
